Rewrite pt2
