+++ app/src/main/java/com/chrisashwalker/menu/Card.java
@@ -1,0 +1,14 @@
+package com.chrisashwalker.menu;
+
+public class Card {
+    static int nextId = 1;
+    int id;
+    String type;
+    int value;
+
+    public Card(String type, int value) {
+        this.id = nextId++;
+        this.type = type;
+        this.value = value;
+    }
+}
